==> UGR/3_UGR_merge_each.ipynb <==
import pandas as pd 
import numpy as np  
from sklearn.utils import shuffle

dt11 = pd.read_csv("../dataset/UGR/traindos.csv")
dt12 = pd.read_csv("../dataset/UGR/generated_dos.csv")
dt21 = pd.read_csv("../dataset/UGR/trainbot.csv")
dt22 = pd.read_csv("../dataset/UGR/generated_bot.csv")

print(dt11['Label'].value_counts())
print(dt21['Label'].value_counts())

a1 = dt12.sample(n=10864, replace=False, random_state=1)
a2 = dt22.sample(n=12599, replace=False, random_state=1)

# Merge data
data1 = pd.concat([dt11,a1],ignore_index = True,axis=0)
data_balanced1 = shuffle(data1).reset_index(drop=True)
#data_balanced1.to_csv("../dataset/UGR/balanced_dos.csv", index=False, header = True)

data2 = pd.concat([dt21,a2],ignore_index = True,axis=0)
data_balanced2 = shuffle(data2).reset_index(drop=True)
data_balanced2.to_csv("../dataset/UGR/balanced_bot.csv", index=False, header = True)


print("Balanced dataset:")
print(data_balanced1['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced2['Label'].value_counts())
print("-----------------------------------------")
