==> UGR/4_merge_full.ipynb <==
# Read original data
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt = pd.read_csv("../dataset/UGR/train_ori.csv")
dt1t = pd.read_csv("../dataset/UGR/test.csv")
dt1a = pd.read_csv("../dataset/UGR/balanced_dos.csv")
dt2a = pd.read_csv("../dataset/UGR/balanced_bot.csv")
dt3a = pd.read_csv("../dataset/UGR/balanced_data_spam.csv")
dt4a = pd.read_csv("../dataset/UGR/balanced_data_scan.csv")

dt1 = dt1a[dt1a['Label']==1]
dt2 = dt2a[dt2a['Label']==1]
dt3 = dt3a[dt3a['Label']==1]
dt4 = dt4a[dt4a['Label']==1]

# Convert label
print("Label converting...")
label = {1: 'dos'}
dt1['Label'] = dt1['Label'].map(label).fillna(dt1['Label'])

label2 = {1: 'nerisbotnet'}
dt2['Label'] = dt2['Label'].map(label2).fillna(dt2['Label'])

label3 = {1: 'anomaly-spam'}
dt3['Label'] = dt3['Label'].map(label3).fillna(dt3['Label'])

label4 = {1: 'anomaly-sshscan'}
dt4['Label'] = dt4['Label'].map(label4).fillna(dt4['Label'])
print('Completed!')

print(dt['Label'].value_counts())
print(dt1t['Label'].value_counts())
print(dt1['Label'].value_counts())
print(dt2['Label'].value_counts())
print(dt3['Label'].value_counts())
print(dt4['Label'].value_counts())

a1 = dt1.sample(n=10864, replace=False, random_state=1)
a2 = dt2.sample(n=12599, replace=False, random_state=1)
a3 = dt3.sample(n=13763, replace=False, random_state=1)
a4 = dt4.sample(n=13954, replace=False, random_state=1)

#New
a1f = dt[dt['Label']=="dos"]
a2f = dt[dt['Label']=="nerisbotnet"]
a3f = dt[dt['Label']=="anomaly-spam"]
a4f = dt[dt['Label']=="anomaly-sshscan"]
a5 = dt[dt['Label']=="blacklist"]
a6 = dt[dt['Label']=="scan11"]
a7 = dt[dt['Label']=="scan44"]
n = dt[dt['Label']=="background"]
n2 = dt1t[dt1t['Label']=="background"]

# New
a80 = n.sample(n=8939, replace=False, random_state=1)
a81 = n2.sample(n=5090, replace=False, random_state=1)

# Merge data
data = pd.concat([a1,a2,a3,a4,a5,a6,a7,a80,a81,a1f,a2f,a3f,a4f],ignore_index = True,axis=0)
data_balanced = shuffle(data).reset_index(drop=True)
data_balanced.to_csv("../dataset/UGR/trainnew.csv", index=False, header = True)
print("Original dataset:")
print(dt['Label'].value_counts())
print("-----------------------------------------")
print("Balanced dataset:")
print(data_balanced['Label'].value_counts())
print("-----------------------------------------")
