==> UGR/4_UGR_Autogluon.ipynb <==
import pandas as pd 
import numpy as np 
import sys
from sklearn.preprocessing import MinMaxScaler 
from autogluon.tabular import TabularDataset, TabularPredictor
import warnings
from fastprogress import fastprogress
warnings.filterwarnings("ignore")

#Load a tabular & test dataset:
train_data = TabularDataset('../dataset/UGR/trainnew.csv')
test_data = TabularDataset('../dataset/UGR/test.csv')
print("Balanced dataset UGR'16:")
print(train_data['Label'].value_counts())
print("Test dataset:")
print(test_data['Label'].value_counts())

print("Fit classification models predicting:")
predictor = TabularPredictor(label='Label', eval_metric='accuracy').fit(train_data) #, presets='best_quality') 

print("The best model:")
print(predictor.get_model_best())
print("------------------------------------------------")
print(predictor.evaluate(test_data))

predictor.fit_summary()

import pandas as pd 
import numpy as np 
import sys
from sklearn.preprocessing import MinMaxScaler 
from autogluon.tabular import TabularDataset, TabularPredictor

#Load a tabular & test dataset:
train_data1 = TabularDataset('../dataset/UGR/train_ori.csv')
test_data1 = TabularDataset('../dataset/UGR/test.csv')

print(train_data1['Label'].value_counts())
print(test_data1['Label'].value_counts())

print("Fit classification models predicting:")
predictor1 = TabularPredictor(label='Label', eval_metric='accuracy').fit(train_data1) 

print("The best model:")
print(predictor1.get_model_best())
print("------------------------------------------------")
print(predictor1.evaluate(test_data1))

print("Evaluate predictions on test data:")
leaderboard1 = predictor1.leaderboard(test_data1)
predictor1.fit_summary()

predictor.feature_importance(data=train_data)
